+++ controllers/DistritoController.php
@@ -8,6 +8,7 @@ use app\models\DistritoSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use app\models\Accesos;
 
 /**
  * DistritoController implements the CRUD actions for Distrito model.
@@ -35,14 +36,27 @@ class DistritoController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new DistritoSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+		
+			if (!\Yii::$app->user->isGuest) 
+			{			
+				if (Accesos::HasAccess(Yii::$app->user->identity->id,"/distrito"))
+				{
+					        $searchModel = new DistritoSearch();
+							$dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
-    }
+							return $this->render('index', [
+								'searchModel' => $searchModel,
+								'dataProvider' => $dataProvider,	
+								]);
+				}else
+				{
+				return $this->redirect(["site/main"]);				
+				}
+			}else
+			{
+				return $this->redirect(["site/login"]);				
+			}
+	}
 
     /**
      * Displays a single Distrito model.
@@ -51,9 +65,24 @@ class DistritoController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+		
+		if (!\Yii::$app->user->isGuest) 
+			{			
+				if (Accesos::HasAccess(Yii::$app->user->identity->id,"/distrito"))
+				{
+					return $this->render('view', [
+					'model' => $this->findModel($id),
+					]);
+				}else
+				{
+				return $this->redirect(["site/main"]);				
+				}
+			}else
+			{
+				return $this->redirect(["site/login"]);				
+			}
+		
+        
     }
 
     /**
@@ -63,15 +92,30 @@ class DistritoController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Distrito();
+		if (!\Yii::$app->user->isGuest) 
+			{			
+				if (Accesos::HasAccess(Yii::$app->user->identity->id,"/distrito"))
+				{
+					      $model = new Distrito();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id_distrito]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
-        }
+						if ($model->load(Yii::$app->request->post()) && $model->save()) {
+							return $this->redirect(['view', 'id' => $model->id_distrito]);
+						} else {
+							return $this->render('create', [
+								'model' => $model,
+							]);
+						}					
+				}else
+				{
+				return $this->redirect(["site/main"]);				
+				}
+			}else
+			{
+				return $this->redirect(["site/login"]);				
+			}
+		
+		
+  
     }
 
     /**
@@ -82,15 +126,28 @@ class DistritoController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+		
+		if (!\Yii::$app->user->isGuest) 
+			{			
+				if (Accesos::HasAccess(Yii::$app->user->identity->id,"/distrito"))
+				{
+					       $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id_distrito]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
-        }
+							if ($model->load(Yii::$app->request->post()) && $model->save()) {
+								return $this->redirect(['view', 'id' => $model->id_distrito]);
+							} else {
+								return $this->render('update', [
+									'model' => $model,
+								]);
+							}					
+				}else
+				{
+				return $this->redirect(["site/main"]);				
+				}
+			}else
+			{
+				return $this->redirect(["site/login"]);				
+			} 
     }
 
     /**
@@ -101,9 +158,22 @@ class DistritoController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+		
+		if (!\Yii::$app->user->isGuest) 
+			{			
+				if (Accesos::HasAccess(Yii::$app->user->identity->id,"/distrito"))
+				{					
+					$this->findModel($id)->delete();
+					return $this->redirect(['index']);
+				}else
+				{
+				return $this->redirect(["site/main"]);				
+				}
+			}else
+			{
+				return $this->redirect(["site/login"]);				
+			}
+		
     }
 
     /**
